Push down ElasticsearchService._generate_paging_dict
