+++ src/azul/service/elasticsearch_service.py
@@ -1,4 +1,3 @@
-import json
 import logging
 from typing import (
     Any,
@@ -59,7 +58,6 @@ from azul.service import (
 )
 from azul.types import (
     JSON,
-    MutableJSON,
 )
 
 log = logging.getLogger(__name__)
@@ -448,63 +446,3 @@ class ElasticsearchService(DocumentService, AbstractService):
         # fetch one more than needed to see if there's a "next page".
         es_search = es_search.extra(size=pagination.size + peek_ahead)
         return es_search
-
-    def _generate_paging_dict(self,
-                              *,
-                              catalog: CatalogName,
-                              filters: FiltersJSON,
-                              es_response: JSON,
-                              pagination: Pagination
-                              ) -> MutableJSON:
-
-        total = es_response['hits']['total']
-        # FIXME: Handle other relations
-        #        https://github.com/DataBiosphere/azul/issues/3770
-        assert total['relation'] == 'eq'
-        pages = -(-total['value'] // pagination.size)
-
-        # ... else use search_after/search_before pagination
-        es_hits = es_response['hits']['hits']
-        count = len(es_hits)
-        if pagination.search_before is None:
-            # hits are normal sorted
-            if count > pagination.size:
-                # There is an extra hit, indicating a next page.
-                count -= 1
-                search_after = tuple(es_hits[count - 1]['sort'])
-            else:
-                # No next page
-                search_after = None
-            if pagination.search_after is not None:
-                search_before = tuple(es_hits[0]['sort'])
-            else:
-                search_before = None
-        else:
-            # hits are reverse sorted
-            if count > pagination.size:
-                # There is an extra hit, indicating a previous page.
-                count -= 1
-                search_before = tuple(es_hits[count - 1]['sort'])
-            else:
-                # No previous page
-                search_before = None
-            search_after = tuple(es_hits[0]['sort'])
-
-        pagination = pagination.advance(search_before=search_before,
-                                        search_after=search_after)
-
-        def page_link(*, previous):
-            return pagination.link(previous=previous,
-                                   catalog=catalog,
-                                   filters=json.dumps(filters))
-
-        return {
-            'count': count,
-            'total': total['value'],
-            'size': pagination.size,
-            'next': page_link(previous=False),
-            'previous': page_link(previous=True),
-            'pages': pages,
-            'sort': pagination.sort,
-            'order': pagination.order
-        }
